Use hostPath instead of PVC
Using PVCs was a good idea, but it does not work out. Each auto-generated PV can be consumed only by one PVC. Creating a PV and PVC makes it more complicated than simply using a hostPath. The bad thing here is that users need to handle with the absolute paths of the shared folders.

Signed-off-by: Volker Theile <votdev@gmx.de>